requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
groq>=0.11.0
python-dotenv